requirements.txt
alembic==1.14.1
asyncpg==0.30.0
bcrypt==4.2.1
fastapi[standart]==0.115.8
ruff==0.9.6
pydantic-settings==2.7.1
pytest==8.3.4
PyJWT==2.10.1
SQLAlchemy==2.0.38